Move code up from ConnectivityService to ClientConnectivityService as most is only relevant for client mode.
Rename SLOW to WARN to make it more generic.

Use NodeNetworkServiceFacade to get num connections.
