+++ shared/domain/src/commonMain/kotlin/network/bisq/mobile/client/websocket/api_proxy/WebSocketApiClient.kt
@@ -13,12 +13,11 @@ import io.ktor.http.HttpStatusCode
 import io.ktor.http.contentType
 import io.ktor.http.isSuccess
 import kotlinx.coroutines.CancellationException
-import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
+import network.bisq.mobile.client.service.network.ClientConnectivityService
 import network.bisq.mobile.client.websocket.WebSocketClientProvider
 import network.bisq.mobile.client.websocket.messages.WebSocketRestApiRequest
 import network.bisq.mobile.client.websocket.messages.WebSocketRestApiResponse
-import network.bisq.mobile.domain.service.network.ConnectivityService
 import network.bisq.mobile.domain.utils.DateUtils
 import network.bisq.mobile.domain.utils.Logging
 import kotlin.uuid.ExperimentalUuidApi
@@ -118,7 +117,7 @@ class WebSocketApiClient(
         try {
             val startTime = DateUtils.now()
             val response = webSocketClientProvider.get().sendRequestAndAwaitResponse(webSocketRestApiRequest)
-            ConnectivityService.newRequestRoundTripTime(DateUtils.now() - startTime)
+            ClientConnectivityService.newRequestRoundTripTime(DateUtils.now() - startTime)
             require(response is WebSocketRestApiResponse) { "Response not of expected type. response=$response" }
             val body = response.body
             if (response.isSuccess()) {
