+++ shared/presentation/src/commonMain/kotlin/network/bisq/mobile/presentation/ui/components/atoms/icons/UserIcon.kt
@@ -1,7 +1,9 @@
 package network.bisq.mobile.presentation.ui.components.atoms.icons
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -36,7 +38,7 @@ fun UserIcon(
 fun GlowEffect(connectivityStatus: ConnectivityService.ConnectivityStatus, modifier: Modifier = Modifier) {
     val name = when (connectivityStatus) {
         ConnectivityService.ConnectivityStatus.CONNECTED -> "green-small-dot.png"
-        ConnectivityService.ConnectivityStatus.SLOW -> "yellow-small-dot.png"
+        ConnectivityService.ConnectivityStatus.WARN -> "yellow-small-dot.png"
         ConnectivityService.ConnectivityStatus.DISCONNECTED -> "red-small-dot.png"
     }
     DynamicImage("drawable/chat/$name", modifier = modifier.scale(2.0F))
